python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymysql==1.1.1
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.0
pydantic==2.9.0
